requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-community
langchain-core
langchain-text-splitters
langchain-google-genai
faiss-cpu
chromadb
pypdf
unstructured
pillow
tiktoken
sentence-transformers
numpy
pandas
google-generativeai==0.8.3
cmake
toml
altair
gitpython
pydeck
watchdog
tornado
blinker
streamlit
protobuf
